Move ssh out of external and into lib
