+++ IPython/parallel/client/client.py
@@ -44,7 +44,7 @@ from IPython.utils.py3compat import cast_bytes, string_types, xrange, iteritems
 from IPython.utils.traitlets import (HasTraits, Integer, Instance, Unicode,
                                     Dict, List, Bool, Set, Any)
 from IPython.external.decorator import decorator
-from IPython.external.ssh import tunnel
+from IPython.lib.ssh import tunnel
 
 from IPython.parallel import Reference
 from IPython.parallel import error
